+++ XSYBBS/app/src/main/java/com/wsg/xsybbs/module/mine/view/MineAdapter.kt
@@ -1,5 +1,6 @@
 package com.wsg.xsybbs.module.mine.view
 
+import cn.bmob.v3.BmobUser
 import com.wsg.xsybbs.module.mine.model.UserSettingInfo
 import me.drakeet.multitype.MultiTypeAdapter
 
@@ -10,6 +11,7 @@ import me.drakeet.multitype.MultiTypeAdapter
 class MineAdapter : MultiTypeAdapter() {
 
     init {
-        register(UserSettingInfo::class.java, UserInfoItemViewBinder())
+//        register(BmobUser::class.java, UserInfoItemViewBinder)
+        register(UserSettingInfo::class.java, UserSettingItemViewBinder())
     }
 }
